Move variant specific static prop logic to domain folder
