+++ packages/app/src/domain/variants/static-props/get-international-variant-table-data.ts
