+++ packages/app/src/domain/variants/static-props/get-variant-chart-data.ts
