+++ packages/app/src/domain/variants/static-props/get-variant-sidebar-value.ts
