+++ packages/app/src/domain/variants/static-props/get-variant-table-data.ts
