+++ packages/app/src/domain/variants/variants-table-tile/components/narrow-variants-table.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 import { isPresent } from 'ts-is-present';
 import { Box } from '~/components/base';
 import { InlineText } from '~/components/typography';
+import { VariantRow } from '~/domain/variants/static-props';
 import { useIntl } from '~/intl';
-import { VariantRow } from '~/static-props/variants/get-variant-table-data';
 import { getMaximumNumberOfDecimals } from '~/utils/get-maximum-number-of-decimals';
 import { useCollapsible } from '~/utils/use-collapsible';
 import {
