+++ packages/app/src/domain/variants/variants-table-tile/components/variants-table.tsx
@@ -1,4 +1,4 @@
-import { VariantRow } from '~/static-props/variants/get-variant-table-data';
+import { VariantRow } from '~/domain/variants/static-props';
 import { useBreakpoints } from '~/utils/use-breakpoints';
 import { TableText } from '../types';
 import { NarrowVariantsTable } from './narrow-variants-table';
